Plutarch.Extra.Interval

Documentation

pmember :: forall a (s :: S). (PEq a, POrd a, PIsData a) => Term s (PAsData a :--> (PInterval a :--> PBool))

pinterval :: forall a (s :: S). PIsData a => Term s (PAsData a :--> (PAsData a :--> PInterval a))

create an interval that includes all values that are greater than or equal - to a and smaller than or equal to b

pfrom :: forall a s. PIsData a => Term s (PAsData a :--> PInterval a)

create an interval that includes all values that are greater than or equal - to a

pto :: forall a (s :: S). PIsData a => Term s (PAsData a :--> PInterval a)

create an interval that includes all values that are smaller than or equal - to a

palways :: forall a (s :: S). (PIsData a, PLiftData a) => Term s (PInterval a)

create an interval that covers every slot

pnever :: forall a (s :: S). (PIsData a, PLiftData a) => Term s (PInterval a)

create an interval that is empty

psingleton :: forall a (s :: S). PIsData a => Term s (PAsData a :--> PInterval a)

create and interval [a, a]

phull :: forall a (s :: S). (PEq a, POrd a, PIsData a) => Term s (PInterval a :--> (PInterval a :--> PInterval a))

`hull i1 i2` is the smallest interval containing i1 and i2

pintersection :: forall a (s :: S). (PEq a, POrd a, PIsData a) => Term s (PInterval a :--> (PInterval a :--> PInterval a))

`intersecion i1 i2` is the largest interval contained in i1 and i2

pcontains :: forall a (s :: S). (PEq a, POrd a, PIsData a) => Term s (PInterval a :--> (PInterval a :--> PBool))

pcontains # a # b is true if the interval b is entirely contained in a

pbefore :: forall a (s :: S). (PEq a, POrd a, PIsData a) => Term s (a :--> (PInterval a :--> PBool))

a before interval i is true if a is earlier than the start of i

pafter :: forall a s. (PEq a, POrd a, PIsData a) => Term s (a :--> (PInterval a :--> PBool))

a after interval i is true if a is later than the end of i